# ReClass Model/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReleaseApprovalRequest": {
        "properties": {
          "serious_discordances": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Serious Discordances",
            "type": "array"
          },
          "signoff_packet": {
            "additionalProperties": true,
            "title": "Signoff Packet",
            "type": "object"
          },
          "target_scope": {
            "additionalProperties": true,
            "title": "Target Scope",
            "type": "object"
          }
        },
        "title": "ReleaseApprovalRequest",
        "type": "object"
      },
      "ReleaseGateRequest": {
        "properties": {
          "active_config_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Config Hash"
          },
          "classification": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Classification"
          },
          "current_state": {
            "default": "review_pending",
            "title": "Current State",
            "type": "string"
          },
          "preflight_failures": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Preflight Failures",
            "type": "array"
          },
          "serious_discordances": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Serious Discordances",
            "type": "array"
          },
          "signoff_packet": {
            "additionalProperties": true,
            "title": "Signoff Packet",
            "type": "object"
          },
          "target_scope": {
            "additionalProperties": true,
            "title": "Target Scope",
            "type": "object"
          }
        },
        "title": "ReleaseGateRequest",
        "type": "object"
      },
      "ReleasePacketRequest": {
        "properties": {
          "benchmark_metrics": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Benchmark Metrics",
            "type": "array"
          },
          "config_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Config Hash"
          },
          "release_scope": {
            "additionalProperties": true,
            "title": "Release Scope",
            "type": "object"
          },
          "serious_discordances": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Serious Discordances",
            "type": "array"
          },
          "sign_off_ledger": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sign Off Ledger",
            "type": "array"
          },
          "source_snapshots": {
            "additionalProperties": true,
            "title": "Source Snapshots",
            "type": "object"
          },
          "validation_report_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Report Id"
          }
        },
        "title": "ReleasePacketRequest",
        "type": "object"
      },
      "ReleaseStateRequest": {
        "properties": {
          "release_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Notes"
          },
          "state": {
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "state"
        ],
        "title": "ReleaseStateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "ValidationRunRequest": {
        "properties": {
          "benchmark": {
            "default": "synthetic_v1",
            "title": "Benchmark",
            "type": "string"
          }
        },
        "title": "ValidationRunRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Tenant-aware service surface over the deterministic ACMG/AMP reclassification engine. Decision support only \u2014 no result is clinically released without credentialed sign-off.",
    "title": "Variant Reclassification API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/validation/release-gate": {
      "post": {
        "operationId": "release_gate_preview_validation_release_gate_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseGateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Release Gate Preview Validation Release Gate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Gate Preview",
        "tags": [
          "validation"
        ]
      }
    },
    "/validation/release-gate/{classification_id}": {
      "get": {
        "operationId": "release_gate_status_validation_release_gate__classification_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "classification_id",
            "required": true,
            "schema": {
              "title": "Classification Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Release Gate Status Validation Release Gate  Classification Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Gate Status",
        "tags": [
          "validation"
        ]
      }
    },
    "/validation/release-gate/{classification_id}/approve": {
      "post": {
        "operationId": "approve_release_validation_release_gate__classification_id__approve_post",
        "parameters": [
          {
            "in": "path",
            "name": "classification_id",
            "required": true,
            "schema": {
              "title": "Classification Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseApprovalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Approve Release Validation Release Gate  Classification Id  Approve Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Approve Release",
        "tags": [
          "validation"
        ]
      }
    },
    "/validation/release-gate/{classification_id}/state": {
      "post": {
        "operationId": "transition_release_state_endpoint_validation_release_gate__classification_id__state_post",
        "parameters": [
          {
            "in": "path",
            "name": "classification_id",
            "required": true,
            "schema": {
              "title": "Classification Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleaseStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Transition Release State Endpoint Validation Release Gate  Classification Id  State Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Transition Release State Endpoint",
        "tags": [
          "validation"
        ]
      }
    },
    "/validation/release-packet": {
      "post": {
        "operationId": "release_packet_preview_validation_release_packet_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReleasePacketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Release Packet Preview Validation Release Packet Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Packet Preview",
        "tags": [
          "validation"
        ]
      }
    },
    "/validation/release-packet/{classification_id}": {
      "get": {
        "operationId": "release_packet_for_classification_validation_release_packet__classification_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "classification_id",
            "required": true,
            "schema": {
              "title": "Classification Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Release Packet For Classification Validation Release Packet  Classification Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Release Packet For Classification",
        "tags": [
          "validation"
        ]
      }
    },
    "/validation/run": {
      "post": {
        "operationId": "run_validation_validation_run_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "X-Tenant-Id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Tenant-Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidationRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Validation Validation Run Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Validation",
        "tags": [
          "validation"
        ]
      }
    },
...
  }
}